requirements.txt
sentence-transformers==3.0.1
faiss-cpu==1.8.0
groq==0.9.0
numpy==1.26.4
pytest==7.4.4
python-dotenv==1.0.0  